packages/agentcore-agent/requirements.txt
bedrock-agentcore
boto3
pyyaml
strands-agents
hypothesis
pytest
